Melhoria na classe de Domínio, service e Controller de Student; Criação dos Dtos e outras alterações;

+++ src/main/java/com/api/register/dto/StudentDto.java
@@ -6,7 +6,7 @@ import com.api.register.enums.Sex;
 import java.time.LocalDate;
 
 public record StudentDto
-        (Integer id, String name, LocalDate age,
+        (Long id, String name, LocalDate age,
          Sex sex, String responsible, String address,
          String cpf, String rg, String birthCertificate) {
     public StudentDto(Student obj){
